Move WorkspaceSearchQuery to workspaces.go
